only track attempts from auto_refresh
