+++ src/node/rpc/node_interface.h
@@ -4,7 +4,6 @@
 
 #include "node/entities.h"
 #include "node/session_metrics.h"
-#include "node/jwt_metrics.h"
 #include "node_call_types.h"
 
 namespace ccf
@@ -51,6 +50,6 @@ namespace ccf
       CodeDigest& code_digest) = 0;
     virtual std::optional<kv::Version> get_startup_snapshot_seqno() = 0;
     virtual SessionMetrics get_session_metrics() = 0;
-    virtual JWTMetrics get_jwt_metrics() = 0;
+    virtual size_t get_jwt_attempts() = 0;
   };
 }
